Moved all the widgets release source to flutter widgets.

+++ packages/syncfusion_pdfviewer_macos/macos/Classes/SyncfusionFlutterPdfViewerPlugin.swift
@@ -22,9 +22,9 @@ public class SyncfusionFlutterPdfViewerPlugin: NSObject, FlutterPlugin {
         {
             initializePdfRenderer(call:call,result:result)
         }
-        else if(call.method == "getImage")
+        else if(call.method == "getPage")
         {
-            getImage(call:call,result:result)
+            getPage(call:call,result:result)
         }
         else if(call.method == "getTileImage")
         {
@@ -48,22 +48,22 @@ public class SyncfusionFlutterPdfViewerPlugin: NSObject, FlutterPlugin {
     private func initializePdfRenderer( call: FlutterMethodCall, result: @escaping FlutterResult)
     {
         guard let argument = call.arguments else {return}
-        let args = argument as? [String: Any]
-        let documentBytes = args!["documentBytes"] as? FlutterStandardTypedData
-        let currentDocumentID = args!["documentID"] as! String
-        let byte = [UInt8](documentBytes!.data)
-        let cfData = CFDataCreate(nil, byte, byte.count)!
-        let dataProvider = CGDataProvider(data: cfData)!
-        let document = CGPDFDocument(dataProvider)
-        documentRepo.updateValue(document!, forKey: currentDocumentID)
-        let pageCount = NSNumber(value: document!.numberOfPages)
+        guard let args = argument as? [String: Any] else {return}
+        guard let documentBytes = args["documentBytes"] as? FlutterStandardTypedData else {return}
+        guard let currentDocumentID = args["documentID"] as? String else {return}
+        let byte = [UInt8](documentBytes.data)
+        guard let cfData = CFDataCreate(nil, byte, byte.count) else {return}
+        guard let dataProvider = CGDataProvider(data: cfData) else {return}
+        guard let document = CGPDFDocument(dataProvider) else {return}
+        documentRepo.updateValue(document, forKey: currentDocumentID)
+        let pageCount = NSNumber(value: document.numberOfPages)
         result(pageCount.stringValue);
     }
     
     private func closeDocument( call: FlutterMethodCall, result: @escaping FlutterResult)
     {
         guard let argument = call.arguments else {return}
-        let documentID = argument as! String
+        guard let documentID = argument as? String  else {return}
         self.documentRepo[documentID] = nil
         self.documentRepo.removeValue(forKey: documentID)
     }
@@ -72,16 +72,16 @@ public class SyncfusionFlutterPdfViewerPlugin: NSObject, FlutterPlugin {
     private func getPagesWidth( call: FlutterMethodCall, result: @escaping FlutterResult)
     {
         guard let argument = call.arguments else {return}
-        let documentID = argument as! String
-        guard let document = self.documentRepo[documentID] else{return}
+        guard let documentID = argument as? String else {return}
+        guard let document = self.documentRepo[documentID] else {return}
         let pageCount = NSNumber(value: document!.numberOfPages)
         var pagesWidth = Array<Double>()
         for index in stride(from: 1,to: pageCount.intValue + 1, by: 1){
-            let page = document!.page(at: Int(index))
-            var pageRect = page!.getBoxRect(.cropBox)
-            if(page!.rotationAngle > 0)
+            guard let page = document!.page(at: Int(index)) else {continue}
+            var pageRect = page.getBoxRect(.cropBox)
+            if(page.rotationAngle > 0)
             {
-                let angle = CGFloat(page!.rotationAngle) * CGFloat.pi/180
+                let angle = CGFloat(page.rotationAngle) * CGFloat.pi/180
                 pageRect = (pageRect.applying(CGAffineTransform(rotationAngle: angle)))
             }
             pagesWidth.append(Double(pageRect.width))
@@ -93,16 +93,16 @@ public class SyncfusionFlutterPdfViewerPlugin: NSObject, FlutterPlugin {
     private func getPagesHeight( call: FlutterMethodCall, result: @escaping FlutterResult)
     {
         guard let argument = call.arguments else {return}
-        let documentID = argument as! String
+        guard let documentID = argument as? String else {return}
         guard let document = self.documentRepo[documentID] else{return}
         let pageCount = NSNumber(value: document!.numberOfPages)
         var pagesHeight = Array<Double>()
         for index in stride(from: 1,to: pageCount.intValue + 1, by: 1){
-            let page = document!.page(at: Int(index))
-            var pageRect = page!.getBoxRect(.cropBox)
-            if(page!.rotationAngle > 0)
+            guard let page = document!.page(at: Int(index)) else {continue}
+            var pageRect = page.getBoxRect(.cropBox)
+            if(page.rotationAngle > 0)
             {
-                let angle = CGFloat(page!.rotationAngle) * CGFloat.pi/180
+                let angle = CGFloat(page.rotationAngle) * CGFloat.pi/180
                 pageRect = (pageRect.applying(CGAffineTransform(rotationAngle: angle)))
             }
             pagesHeight.append(Double(pageRect.height))
@@ -110,117 +110,117 @@ public class SyncfusionFlutterPdfViewerPlugin: NSObject, FlutterPlugin {
         result(pagesHeight)
     }
     
-    // Gets the pdf page image from the specified page
-    private func getImage( call: FlutterMethodCall, result: @escaping FlutterResult)
+    // Gets the image bytes of the specified page from the document at the specified width and height
+    private func getPage( call: FlutterMethodCall, result: @escaping FlutterResult)
     {
         guard let argument = call.arguments else {return}
-        let args = argument as? [String: Any]
-        let index = args!["index"] as? Int
-        var scale = CGFloat(args!["scale"] as! Double)
-        if(scale < 2)
-        {
-            scale = 2
-        }
-        let documentID = args!["documentID"] as! String
-        guard let image = getImageForPlugin(index: index!,scale: scale,documentID: documentID) else {
-            result(FlutterStandardTypedData())
+        guard let args = argument as? [String: Any] else {return}
+        guard let index = args["index"] as? Int else {return}
+        guard let width = args["width"] as? Int else {return}
+        guard let height = args["height"] as? Int else {return}
+        guard let documentID = args["documentID"] as? String else {return}
+        guard let image = getImageForPlugin(index: index, width: width, height: height,documentID: documentID) else {
             return
         }
         result(image)
     }
     
     // Gets the image for plugin
-    private func getImageForPlugin(index: Int,scale: CGFloat,documentID: String) -> FlutterStandardTypedData?
+    private func getImageForPlugin(index: Int,width:Int, height:Int,documentID: String) -> FlutterStandardTypedData?
     {
         guard let document = self.documentRepo[documentID] else {return nil}
-        let page = document!.page(at: Int(index))
-        let pageRect = page!.getBoxRect(.cropBox)
-        let imageRect = CGRect(x: 0,y: 0,width: pageRect.size.width*CGFloat(scale),height: pageRect.size.height*CGFloat(scale))
-        let nsImage = NSImage(size: imageRect.size, actions: { cgContext in
-            let transform = page!.getDrawingTransform(.cropBox, rect: CGRect(origin: CGPoint.zero, size: pageRect.size), rotate: 0, preserveAspectRatio: true)
-            cgContext.translateBy(x: 0.0, y: imageRect.size.height)
-            cgContext.scaleBy(x: CGFloat(scale), y: -CGFloat(scale))
-            cgContext.concatenate(transform)
-            cgContext.drawPDFPage(page!)
-            cgContext.endPage()
-        })
-        let bytes = nsImage.tiffRepresentation?.bitmap?.png
-        return bytes == nil ? FlutterStandardTypedData() : FlutterStandardTypedData(bytes: bytes!)
+        guard let page = document!.page(at: Int(index)) else {return nil}
+        let pageRect = page.getBoxRect(.cropBox)
+        var pageWidth = pageRect.width
+        var pageHeight = pageRect.height
+        if(page.rotationAngle == 90 || page.rotationAngle == 270) {
+            pageWidth = pageRect.height
+            pageHeight = pageRect.width
+        }
+        let scaleX = Double(width) / Double(pageWidth)
+        let scaleY = Double(height) / Double(pageHeight)
+        let stride = width * 4
+        let bufSize = stride * height;
+        let buffer = UnsafeMutablePointer<UInt8>.allocate(capacity: bufSize)
+        buffer.initialize(repeating:  0, count: bufSize)
+        var rendered = false
+        let transform = page.getDrawingTransform(.cropBox, rect: CGRect(origin: CGPoint.zero, size: CGSize(width: pageWidth, height: pageHeight)), rotate: 0, preserveAspectRatio: true)
+        let rgb = CGColorSpaceCreateDeviceRGB()
+        let context = CGContext(data: buffer, width: width, height: height, bitsPerComponent: 8, bytesPerRow: stride, space: rgb, bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue)
+        if context != nil {
+            context!.setAllowsAntialiasing(true);
+            context!.scaleBy(x: scaleX, y: scaleY)
+            context!.concatenate(transform)
+            context!.drawPDFPage(page)
+            rendered = true
+        }
+        if(rendered){
+            let data = Data(bytesNoCopy: buffer, count: bufSize, deallocator: .none)
+            return FlutterStandardTypedData(bytes: data)
+        }else{
+            return nil
+        }
     }
     
     // Gets the pdf page image from the specified page
     private func getTileImage( call: FlutterMethodCall, result: @escaping FlutterResult)
     {
         guard let argument = call.arguments else {return}
-        let args = argument as? [String: Any]
-        let pageNumber = args!["pageNumber"] as? Int
-        let scale = CGFloat(args!["scale"] as! Double)
-        let width = args!["width"] as! Double
-        let height = args!["height"] as! Double
-        let x = args!["x"] as! Double
-        let y = args!["y"] as! Double
- 
-        let documentID = args!["documentID"] as! String
-        guard let tileImage = getTileImageForPlugin(pageNumber: pageNumber!, scale: scale,
-              width: width, height: height, x: x, y: y, documentID: documentID)
+        guard let args = argument as? [String: Any] else {return}
+        guard let pageNumber = args["pageNumber"] as? Int else {return}
+        guard let scale = args["scale"] as? Double else {return}
+        guard let width = args["width"] as? Double else {return}
+        guard let height = args["height"] as? Double else {return}
+        guard let x = args["x"] as? Double else {return}
+        guard let y = args["y"] as? Double else {return}
+        guard let documentID = args["documentID"] as? String else {return}
+        guard let tileImage = getTileImageForPlugin(pageNumber: pageNumber, scale: CGFloat(scale),
+                                                    width: width, height: height, x: x, y: y, documentID: documentID)
         else {
-            result(FlutterStandardTypedData())
             return
-        }      
+        }
         result(tileImage)
     }
-        
+    
     // Gets the image for plugin
     private func getTileImageForPlugin(pageNumber: Int, scale: CGFloat, width: Double, height: Double, x: Double, y: Double, documentID: String) -> FlutterStandardTypedData?
     {
         guard let document = self.documentRepo[documentID] else {return nil}
-        let page = document!.page(at: Int(pageNumber))
-        let pageRect = page!.getBoxRect(.cropBox)
-        
+        guard let page = document!.page(at: Int(pageNumber)) else {return nil}
+        let pageRect = page.getBoxRect(.cropBox)
         var pageWidth = pageRect.width
         var pageHeight = pageRect.height
-        
-        if(page!.rotationAngle == 90 || page!.rotationAngle == 270) {
+        if(page.rotationAngle == 90 || page.rotationAngle == 270) {
             pageWidth = pageRect.height
             pageHeight = pageRect.width
         }
-        let imageRect = CGRect(x: 0,y: 0, width: width, height: height)
         let bounds = CGRect(x: -(pageWidth * scale / 2) + (pageWidth / 2) - CGFloat(x),
                             y: -(pageHeight * scale / 2) + (pageHeight / 2) + CGFloat(y),
                             width: pageWidth * scale, height: pageHeight * scale)
-        let nsImage = NSImage(size: imageRect.size, actions: { cgContext in
-            let transform = page!.getDrawingTransform(.cropBox, rect: bounds, rotate: 0, preserveAspectRatio: true)
-            cgContext.translateBy(x: 0.0, y: pageHeight * scale)
-            cgContext.scaleBy(x: scale, y: -scale)
-            cgContext.setFillColor(red: 1.0, green: 1.0, blue: 1.0, alpha: 1.0)
-            cgContext.fill(bounds)
-            cgContext.concatenate(transform)
-            cgContext.drawPDFPage(page!)
-            cgContext.endPage()
-        })
-        let bytes = nsImage.tiffRepresentation?.bitmap?.png
-        return bytes == nil ? FlutterStandardTypedData() : FlutterStandardTypedData(bytes: bytes!)
+        
+        let stride = Int(width) * 4
+        let bufSize = stride * Int(height);
+        let buffer = UnsafeMutablePointer<UInt8>.allocate(capacity: bufSize)
+        buffer.initialize(repeating:  0, count: bufSize)
+        var rendered = false
+        let transform = page.getDrawingTransform(.cropBox, rect: CGRect(origin: CGPoint.zero, size: CGSize(width: pageWidth, height: pageHeight)), rotate: 0, preserveAspectRatio: true)
+        let rgb = CGColorSpaceCreateDeviceRGB()
+        let context = CGContext(data: buffer, width: Int(width), height: Int(height), bitsPerComponent: 8, bytesPerRow: stride, space: rgb, bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue)
+        if context != nil {
+            context!.setAllowsAntialiasing(true);
+            context!.translateBy(x: CGFloat(-x * scale), y: CGFloat(((y * scale) + height) - bounds.height))
+            context!.scaleBy(x: scale, y: scale)
+            context!.setFillColor(red: 1.0, green: 1.0, blue: 1.0, alpha: 1.0)
+            context!.fill(bounds)
+            context!.concatenate(transform)
+            context!.drawPDFPage(page)
+            rendered = true
+        }
+        if(rendered){
+            let data = Data(bytesNoCopy: buffer, count: bufSize, deallocator: .none)
+            return FlutterStandardTypedData(bytes: data)
+        }else{
+            return nil
+        }
     }
-}
-
-extension NSBitmapImageRep {
-    var png: Data? {
-        return representation(using: .png, properties: [:])
-    }
-}
-
-extension Data {
-    var bitmap: NSBitmapImageRep? {
-        return NSBitmapImageRep(data: self)
-    }
-}
-
-extension NSImage {
-    convenience init(size: CGSize, actions: (CGContext) -> Void) {
-        self.init(size: size)
-        lockFocusFlipped(true)
-        actions(NSGraphicsContext.current!.cgContext)
-        unlockFocus()
-    }
-    var png: Data? { tiffRepresentation?.bitmap?.png }
 }
